running state
