+++ criteo_distributed_training/autoscalar_aws.yml
@@ -3,7 +3,7 @@ cluster_name: default
 
 # The maximum number of workers nodes to launch in addition to the head
 # node.
-max_workers: 47
+max_workers: 11
 
 # The autoscaler will scale up the cluster faster with higher upscaling speed.
 # E.g., if the task requires adding more nodes then autoscaler will gradually
@@ -52,7 +52,7 @@ auth:
 # By default Ray creates a new private keypair, but you can also use your own.
 # If you do so, make sure to also set "KeyName" in the head and worker node
 # configurations below.
-#    ssh_private_key: /path/to/your/key.pem
+    ssh_private_key: /home/mritunjay/MyEC2KeyPair.pem
 
 # Tell the autoscaler the allowed node types and the resources they provide.
 # The key is the name of the node type, which is just for debugging purposes.
@@ -64,7 +64,7 @@ available_node_types:
         # You can also set custom resources.
         # For example, to mark a node type as having 1 CPU, 1 GPU, and 5 units of a resource called "custom", set
         # resources: {"CPU": 1, "GPU": 1, "custom": 5}
-        resources: {"CPU": 4}
+        resources: {"CPU": 6}
         # Provider-specific config for this node type, e.g. instance type. By default
         # Ray will auto-configure unspecified fields such as SubnetId and KeyName.
         # For more documentation on available fields, see:
@@ -78,14 +78,16 @@ available_node_types:
                 - DeviceName: /dev/sda1
                   Ebs:
                       VolumeSize: 40
+            KeyName: MyEC2KeyPair
+            
             # Additional options in the boto docs.
     ray.worker.default:
         # The minimum number of worker nodes of this type to launch.
         # This number should be >= 0.
-        min_workers: 47
+        min_workers: 11
         # The maximum number of worker nodes of this type to launch.
         # This takes precedence over min_workers.
-        max_workers: 47
+        max_workers: 11
         # The node type's CPU and GPU resources are auto-detected based on AWS instance type.
         # If desired, you can override the autodetected CPU and GPU resources advertised to the autoscaler.
         # You can also set custom resources.
@@ -103,6 +105,10 @@ available_node_types:
                 - DeviceName: /dev/sda1
                   Ebs:
                       VolumeSize: 25
+            KeyName: MyEC2KeyPair
+            IamInstanceProfile:
+                Arn: arn:aws:iam::199696198976:instance-profile/ray-autoscaler-v1
+
             # Run workers on spot by default. Comment this out to use on-demand.
             # NOTE: If relying on spot instances, it is best to specify multiple different instance
             # types to avoid interruption when one instance type is experiencing heightened demand.
@@ -119,9 +125,8 @@ head_node_type: ray.head.default
 
 # Files or directories to copy to the head and worker nodes. The format is a
 # dictionary from REMOTE_PATH: LOCAL_PATH, e.g.
-file_mounts: {
-#    "/path1/on/remote/machine": "/path1/on/local/machine",
-#    "/path2/on/remote/machine": "/path2/on/local/machine",
+file_mounts: { 
+    "/home/ubuntu/criteo_distributed_training": "/home/mritunjay/Public-Benchmarks/criteo_distributed_training"
 }
 
 # Files or directories to copy from the head node to the worker nodes. The format is a
@@ -154,7 +159,8 @@ initialization_commands: []
 setup_commands: 
   - sudo apt-get update -y
   - sudo DEBIAN_FRONTEND=noninteractive -E apt install python3-pip -y
-  - pip3 install thirdai 'ray[default]' botocore boto3 https://github.com/pratkpranav/pygloo/releases/download/0.2.0/pygloo-0.2.0-cp310-cp310-linux_x86_64.whl numpy==1.23.5 s3fs "pydantic<2" torch
+  - sudo DEBIAN_FRONTEND=noninteractive -E apt install awscli -y
+  - pip3 install thirdai 'ray[default]' botocore boto3 pyarrow https://github.com/pratkpranav/pygloo/releases/download/0.2.0/pygloo-0.2.0-cp310-cp310-linux_x86_64.whl numpy==1.23.5 s3fs "pydantic<2" torch
   - export PATH=$PATH:/home/$USER/.local/bin
     # Note: if you're developing Ray, you probably want to create a Docker image that
     # has your Ray repo pre-cloned. Then, you can replace the pip installs
@@ -165,8 +171,8 @@ setup_commands:
 
 # Custom commands that will be run on the head node after common setup.
 head_setup_commands:
-  - pip3 install scikit-learn
-  - git clone https://github.com/ThirdAILabs/Public-Benchmarks.git
+    - pip3 install scikit-learn
+#   - git clone https://github.com/ThirdAILabs/Public-Benchmarks.git
 
 # Custom commands that will be run on worker nodes after common setup.
 worker_setup_commands: []
@@ -174,7 +180,7 @@ worker_setup_commands: []
 # Command to start ray on the head node. You don't need to change this.
 head_start_ray_commands:
     - ray stop
-    - ray start --head --port=6379 --object-manager-port=8076 --autoscaling-config=~/ray_bootstrap_config.yaml
+    - ray start --head --port=6379 --object-manager-port=8076 --autoscaling-config=~/ray_bootstrap_config.yaml --system-config='{"core_worker_internal_heartbeat_ms":10000}'
 
 # Command to start ray on worker nodes. You don't need to change this.
 worker_start_ray_commands:
